+++ RGBA Synth/Source/PluginProcessor.h
@@ -4,6 +4,8 @@
 
 #include <JuceHeader.h>
 #include "RGBASynthesizer.h"
+#include "RGBASynthSounds.h"
+#include "RGBASynthVoices.h"
 
 
 //==============================================================================
@@ -106,6 +108,8 @@ private:
     int extraVoices;
 
     RGBASynthesizer RGBASynth;
+    RGBASound1 RGBAsnd1;
+    RGBASin RGBASinVoice;
 
     juce::MidiMessageCollector midiCollector;
